+++ CHANGELOG.md
@@ -192,6 +192,8 @@
 + kvs::EventListener::enableEvent( event_type )
 + kvs::EventListener::disableEvent()
 + kvs::EventListener::disableEvent( event_type )
++ kvs::mpi::Communicator::abort()
++ kvs::mpi::Communicator::barrier()
 
 **Added new examples**
 + Example/Visualization/ScatterPlotMatrixRenderer